backend-python/requirements.txt
--index-url https://pypi.tuna.tsinghua.edu.cn/simple/

# Web框架核心
fastapi==0.128.0
uvicorn==0.40.0
starlette==0.50.0
sse-starlette==3.2.0
python-multipart==0.0.22

# 数据验证/配置
pydantic==2.12.5
PyYAML==6.0.3

# 文档解析
pypdf==6.6.2
docx2txt==0.9
PyPDF2==3.0.1
python-docx==1.2.0
markdown==3.10.1
beautifulsoup4==4.14.3
unstructured==0.10.30

# 向量数据库
chromadb==1.4.1
langchain-chroma==1.1.0

# LangChain
langchain==1.2.7
langchain-community==0.4.1
langchain-core==1.2.7
langchain-text-splitters==1.1.0

# AI
dashscope==1.25.10
requests==2.32.5
aiohttp==3.13.3

# Agent
tavily-python==0.7.20

# 轻量级图像处理
onnxruntime==1.20.1
opencv-python-headless==4.11.0.86
numpy==1.26.4
rembg==2.0.60
pillow==10.3.0